+++ src/precompile.jl
@@ -11,7 +11,7 @@ using PrecompileTools
     let
         x = [0.1, 1.1, 2.1, 3.1, 4.1]
         domain = Domain(x)
-        model = Model(@λ (x, a2=1, a1=1, a0=5) -> (a2 .* x.^2  .+  a1 .* x  .+  a0))
+        model = Model(@fd (x, a2=1, a1=1, a0=5) -> (a2 .* x.^2  .+  a1 .* x  .+  a0))
         data = GModelFit.mock(Measures, model, domain, seed=1)
         status = fit(model, data)
 
@@ -22,7 +22,7 @@ using PrecompileTools
                       :bkg => GModelFit.OffsetSlope(0.5, 1, 0.1),
                       :main => SumReducer(:l1, :l2, :bkg));
         model[:l2].norm.patch = :l1
-        model[:l2].norm.patch = @λ (m, v) -> v + m[:l1].norm
+        model[:l2].norm.patch = @fd (m, v) -> v + m[:l1].norm
         data = GModelFit.mock(Measures, model, Domain(x), seed=1)
         status = fit(model, data, minimizer=cmpfit())
 
@@ -44,9 +44,9 @@ using PrecompileTools
         thaw!(model[1], :bkg);
         thaw!(model[2], :bkg);
 
-        model[2][:bkg].offset.mpatch = @λ m -> m[1][:bkg].offset
-        model[2][:bkg].slope.mpatch  = @λ m -> m[1][:bkg].slope
-        model[1][:l2].center.mpatch = @λ m -> m[2][:l2].center
+        model[2][:bkg].offset.mpatch = @fd m -> m[1][:bkg].offset
+        model[2][:bkg].slope.mpatch  = @fd m -> m[1][:bkg].slope
+        model[1][:l2].center.mpatch = @fd m -> m[2][:l2].center
 
         data = GModelFit.mock(Measures, model, [Domain(x), Domain(x)], seed=1)
         status = fit(model, data)